feat: handle image uploads (#4)
## Demo

https://github.com/user-attachments/assets/9ccab570-ed6e-4223-92a0-ac039cb373ee

## Description

In the demo video, there's an issue where updates are not happening
immediately. In reality, there's an intermittent issue on iOS where the
Postgres subscription set up in Supabase gets disconnected.

Additionally, please add the functionality to upload multiple images in
posts and replies. For posts, ensure that the images are recognized
correctly and allow users to change images if needed.